routing progress, stuck on a weird home bug that i think has to do with asynchronous functions/calls
