+++ client/Signup.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Signup = () => {
+  return <div>Signup</div>
+}
+
+export default Signup
