+++ package.json
@@ -60,7 +60,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-redux": "^7.0.1",
-    "react-router-dom": "^5.0.0",
+    "react-router-dom": "^5.2.0",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
