+++ server/auth/reddit.js
@@ -42,8 +42,8 @@ router.get('/callback', (req, res, next) => {
   // Check for origin via state token
   if (req.query.state === req.session.state) {
     passport.authenticate('reddit', {
-      failureRedirect: '/login',
-      successRedirect: '/'
+      failureRedirect: '/',
+      successRedirect: '/home'
     })(req, res, next)
   } else {
     next(new Error(403))
